chore: Move ids.ts to mirror-server
The IDs are only created on the server and thus `friendly-words` is only
needed on the server.
