+++ mirror/mirror-server/src/ids.ts
@@ -1,6 +1,6 @@
 import {objects, predicates} from 'friendly-words';
-import * as base62 from '../base62.js';
-import * as crypto from '../crypto.js';
+import * as base62 from 'shared/src/base62.js';
+import * as crypto from 'shared/src/crypto.js';
 
 const tempUint64Array = new BigUint64Array(1);
