+++ packages/shared/package.json
@@ -14,7 +14,6 @@
   "devDependencies": {
     "@rocicorp/eslint-config": "^0.4.2",
     "@rocicorp/prettier-config": "^0.1.1",
-    "@types/friendly-words": "^1.2.0",
     "ts-jest": "^29.1.0",
     "typescript": "^5.1.3"
   },
@@ -22,7 +21,6 @@
     "@badrap/valita": "^0.2.10",
     "@rocicorp/logger": "^5.2.0",
     "@rocicorp/resolver": "^1.0.0",
-    "friendly-words": "^1.2.3",
     "semver": "^7.5.3"
   },
   "eslintConfig": {
